+++ internal/batches/ui/json_lines.go
@@ -177,6 +177,14 @@ func (ui *JSONLines) WriteAfterStepResult(key string, value execution.AfterStepR
 	})
 }
 
+func (ui *JSONLines) DockerWatchDogWarning(err error) {
+	message := fmt.Sprintf(`It seems your Docker engine might be frozen.
+If there's no progress in the next couple minutes, you may want to try restarting Docker and running the command again.
+Error: %s
+`, err.Error())
+	logOperationFailure(batcheslib.LogEventOperationDockerWatchDog, &batcheslib.DockerWatchDogMetadata{Error: message})
+}
+
 type taskExecutionJSONLines struct {
 	linesTasks  map[*executor.Task]batcheslib.JSONLinesTask
 	binaryDiffs bool